`Pipeline` refactor to use `BasePipeline` (#1081)
* initial refactor

* update test; finish off initial refactoring changes post local testing

* test fix

* Test BasePipeline register and create

* add tests for BasePipeline

* add and update docstring

* update test

* move BasePipeline to a new file

* test fix

* anothe test fix

* fix import

* revert

* rebase fix

* expose functions through pipeline.py
