+++ tests/deepsparse/image_classification/test_pipelines.py
@@ -56,7 +56,7 @@ def test_image_classification_pipeline_preprocessing(
         ]
     )
 
-    ic_pipeline = Pipeline.create("image_classification", zoo_stub)
+    ic_pipeline = Pipeline.create("image_classification", model_path=zoo_stub)
     zoo_model = Model(zoo_stub)
     data_originals_path = None
     if zoo_model.sample_originals is not None:
